requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.4
GeoAlchemy2==0.15.2
h11==0.14.0
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
numpy>=1.20.0  # Required dependency for scipy
packaging==24.1
pandas>=1.3.0     # Required dependency
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scipy>=1.7.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
folium>=0.12.0
matplotlib>=3.5.0
seaborn>=0.11.0
matplotlib>=3.4.0  # Required dependency
scikit-learn>=1.0.0
